Extract gzip and digest helper classes to top level
