+++ textrepo-app/src/main/java/nl/knaw/huc/helpers/gzip/GzipHelper.java
@@ -1,4 +1,4 @@
-package nl.knaw.huc.helpers;
+package nl.knaw.huc.helpers.gzip;
 
 import java.util.zip.GZIPInputStream;
